requirements.txt
# ESG Analyzer v2.3 - Requirements

# Core
streamlit>=1.32.0
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0
pandas>=2.0.0

# LLM
litellm>=1.30.0
openai>=1.12.0
anthropic>=0.18.0

# Document processing
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.1.0

# PDF processing
PyMuPDF>=1.23.0

# HTTP
requests>=2.31.0
tenacity>=8.2.0

# Utilities
tiktoken>=0.5.0
